feat(core_config):support la464 and centaur320 core
