+++ main.c
@@ -477,11 +477,6 @@ void cpu_reset(CPUState* cs) {
     env->CSR_MERRCTL = FIELD_DP64(env->CSR_MERRCTL, CSR_MERRCTL, ISMERR, 0);
     env->CSR_TID = cs->cpu_index;
 
-    env->CSR_PRCFG3 = FIELD_DP64(env->CSR_PRCFG3, CSR_PRCFG3, TLB_TYPE, 2);
-    env->CSR_PRCFG3 = FIELD_DP64(env->CSR_PRCFG3, CSR_PRCFG3, MTLB_ENTRY, 63);
-    env->CSR_PRCFG3 = FIELD_DP64(env->CSR_PRCFG3, CSR_PRCFG3, STLB_WAYS, 7);
-    env->CSR_PRCFG3 = FIELD_DP64(env->CSR_PRCFG3, CSR_PRCFG3, STLB_SETS, 8);
-
     for (n = 0; n < 4; n++) {
         env->CSR_DMW[n] = FIELD_DP64(env->CSR_DMW[n], CSR_DMW, PLV0, 0);
         env->CSR_DMW[n] = FIELD_DP64(env->CSR_DMW[n], CSR_DMW, PLV1, 0);
@@ -501,87 +496,6 @@ void cpu_reset(CPUState* cs) {
     restore_fp_status(env);
 #endif
     cs->exception_index = -1;
-}
-
-void loongarch_la464_initfn(CPULoongArchState* env) {
-    int i;
-
-    for (i = 0; i < 21; i++) {
-        env->cpucfg[i] = 0x0;
-    }
-
-    env->cpucfg[0] = 0x14c010;  /* PRID */
-
-    uint32_t data = 0;
-    data = FIELD_DP32(data, CPUCFG1, ARCH, 2);
-    data = FIELD_DP32(data, CPUCFG1, PGMMU, 1);
-    data = FIELD_DP32(data, CPUCFG1, IOCSR, 1);
-    data = FIELD_DP32(data, CPUCFG1, PALEN, 0x2f);
-    data = FIELD_DP32(data, CPUCFG1, VALEN, 0x2f);
-    data = FIELD_DP32(data, CPUCFG1, UAL, 1);
-    data = FIELD_DP32(data, CPUCFG1, RI, 1);
-    data = FIELD_DP32(data, CPUCFG1, EP, 1);
-    data = FIELD_DP32(data, CPUCFG1, RPLV, 1);
-    data = FIELD_DP32(data, CPUCFG1, HP, 1);
-    data = FIELD_DP32(data, CPUCFG1, IOCSR_BRD, 1);
-    env->cpucfg[1] = data;
-
-    data = 0;
-    data = FIELD_DP32(data, CPUCFG2, FP, 1);
-    data = FIELD_DP32(data, CPUCFG2, FP_SP, 1);
-    data = FIELD_DP32(data, CPUCFG2, FP_DP, 1);
-    data = FIELD_DP32(data, CPUCFG2, FP_VER, 1);
-    data = FIELD_DP32(data, CPUCFG2, LSX, 1),
-    data = FIELD_DP32(data, CPUCFG2, LASX, 1),
-    data = FIELD_DP32(data, CPUCFG2, LLFTP, 1);
-    data = FIELD_DP32(data, CPUCFG2, LLFTP_VER, 1);
-    data = FIELD_DP32(data, CPUCFG2, LSPW, 1);
-    data = FIELD_DP32(data, CPUCFG2, LAM, 1);
-    env->cpucfg[2] = data;
-
-    env->cpucfg[4] = 100 * 1000 * 1000; /* Crystal frequency */
-
-    data = 0;
-    data = FIELD_DP32(data, CPUCFG5, CC_MUL, 1);
-    data = FIELD_DP32(data, CPUCFG5, CC_DIV, 1);
-    env->cpucfg[5] = data;
-
-    data = 0;
-    data = FIELD_DP32(data, CPUCFG16, L1_IUPRE, 1);
-    data = FIELD_DP32(data, CPUCFG16, L1_DPRE, 1);
-    data = FIELD_DP32(data, CPUCFG16, L2_IUPRE, 1);
-    data = FIELD_DP32(data, CPUCFG16, L2_IUUNIFY, 1);
-    data = FIELD_DP32(data, CPUCFG16, L2_IUPRIV, 1);
-    data = FIELD_DP32(data, CPUCFG16, L3_IUPRE, 1);
-    data = FIELD_DP32(data, CPUCFG16, L3_IUUNIFY, 1);
-    data = FIELD_DP32(data, CPUCFG16, L3_IUINCL, 1);
-    env->cpucfg[16] = data;
-
-    data = 0;
-    data = FIELD_DP32(data, CPUCFG17, L1IU_WAYS, 3);
-    data = FIELD_DP32(data, CPUCFG17, L1IU_SETS, 8);
-    data = FIELD_DP32(data, CPUCFG17, L1IU_SIZE, 6);
-    env->cpucfg[17] = data;
-
-    data = 0;
-    data = FIELD_DP32(data, CPUCFG18, L1D_WAYS, 3);
-    data = FIELD_DP32(data, CPUCFG18, L1D_SETS, 8);
-    data = FIELD_DP32(data, CPUCFG18, L1D_SIZE, 6);
-    env->cpucfg[18] = data;
-
-    data = 0;
-    data = FIELD_DP32(data, CPUCFG19, L2IU_WAYS, 15);
-    data = FIELD_DP32(data, CPUCFG19, L2IU_SETS, 8);
-    data = FIELD_DP32(data, CPUCFG19, L2IU_SIZE, 6);
-    env->cpucfg[19] = data;
-
-    data = 0;
-    data = FIELD_DP32(data, CPUCFG20, L3IU_WAYS, 15);
-    data = FIELD_DP32(data, CPUCFG20, L3IU_SETS, 14);
-    data = FIELD_DP32(data, CPUCFG20, L3IU_SIZE, 6);
-    env->cpucfg[20] = data;
-
-    env->CSR_ASID = FIELD_DP64(0, CSR_ASID, ASIDBITS, 0xa);
 }
 
 static void loongarch_cpu_do_interrupt(CPUState *cs)
@@ -1157,7 +1071,7 @@ int main(int argc, char** argv, char **envp) {
     CPULoongArchState* env = &cpu->env;
     cs->env = env;
     cpu_reset(cs);
-    loongarch_la464_initfn(env);
+    loongarch_core_initfn(env);
     cpu_clear_tc(env);
     env->timer_counter = INT64_MAX;
 #ifndef CONFIG_USER_ONLY
